+++ src/text-views/blockquote.js
@@ -2,7 +2,11 @@ import usage from './blockquote.usage.js';
 
 export default function({ textView }) {
     textView.define('blockquote', function(node, props, data, context) {
-        const { content = 'text' } = props;
+        const { content = 'text', kind } = props;
+
+        if (typeof kind === 'string') {
+            node.appendLine().appendText(`[!${kind.toUpperCase()}]`);
+        }
 
         return textView.render(node, content, data, context);
     }, {