+++ src/text-views/blockquote.usage.js
@@ -5,5 +5,12 @@ export default (view) => ({
             'text:"Some text"',
             { view, content: 'text:"Nested blockquote\\nmulti-line text"' }
         ]
-    }
+    },
+    examples: [
+        {
+            title: 'Variations',
+            demoData: 'Blockquote text...',
+            demo: ['note', 'tip', 'important', 'warning', 'caution'].map(name => `blockquote{ kind: "${name}" }`)
+        }
+    ]
 });