+++ src/views/text/blockquote.usage.js
@@ -2,5 +2,12 @@ export default (view) => ({
     demo: {
         view,
         data: '"Blockquote text"'
-    }
+    },
+    examples: [
+        {
+            title: 'Variations',
+            demoData: 'Blockquote text...',
+            demo: ['note', 'tip', 'important', 'warning', 'caution'].map(name => `blockquote{ kind: "${name}" }`)
+        }
+    ]
 });